Use CHICKEN tools from the current runtime system
Generate the makefile defaulting to calls to the CHICKEN tools from
the runtime system which executed egg-layer.

The CHICKEN tools and CHICKEN_REPOSITORY_PATH can be tweaked from the
environment if the Makefile is to be processed on a system without
egg-layer.  The following environment variables can be used for that:

* CSI
* CHICKEN_INSTALL
* CHICKEN_REPOSITORY_PATH

Also add two new parameters (egg-layer-params):

* csi-program
* chicken-install-program